Removed duplicated path normalization logic by introducing shared utilities and refactoring both call sites to use them. Signed-off-by: Corwin Joy <corwin.joy@gmail.com>
